move shim to validate.go
